+++ app/lib/constants.js
@@ -5,6 +5,7 @@ module.exports = {
   promHistogramBuckets: [0.01, 0.05, 0.1, 0.2, 0.3, 0.5, 1, 1.5, 5, 10],
   promEsQueryLabelName: 'query_type',
   searchTypes: {
+    kits16to24: 'kits_16to24_geo',
     sexperts: 'sexual_health_professionals_geo',
   },
   symptoms: {